requirements.txt
pkg_resources==0.0.0
pycrypto==2.6.1
pycryptodome==3.13.0
